requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.6.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=pypi_0
autopep8=1.6.0=pyhd3eb1b0_1
aws-cdk-aws-lambda-python-alpha=2.46.0a0=pypi_0
aws-cdk-lib=2.46.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bzip2=1.0.8=he774522_0
ca-certificates=2022.07.19=haa95532_0
cattrs=22.1.0=pypi_0
certifi=2022.9.24=py310haa95532_0
charset-normalizer=2.1.1=pypi_0
click=8.0.4=py310haa95532_0
colorama=0.4.5=py310haa95532_0
constructs=10.1.131=pypi_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
docopt=0.6.2=pypi_0
entrypoints=0.4=py310haa95532_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.0.0rc9=pypi_0
executing=0.8.3=pyhd3eb1b0_0
h11=0.12.0=pypi_0
httpcore=0.15.0=pypi_0
httpx=0.23.0=pypi_0
idna=3.4=pypi_0
iniconfig=1.1.1=pypi_0
ipykernel=6.15.2=py310haa95532_0
ipython=8.4.0=py310haa95532_0
jedi=0.18.1=py310haa95532_1
jsii=1.69.0=pypi_0
jupyter_client=7.3.5=py310haa95532_0
jupyter_core=4.11.1=py310haa95532_0
libffi=3.4.2=hd77b12b_4
libsodium=1.0.18=h62dcd97_0
matplotlib-inline=0.1.6=py310haa95532_0
nest-asyncio=1.5.5=py310haa95532_0
numpy=1.23.4=pypi_0
openpyxl=3.0.10=pypi_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=22.2.2=py310haa95532_0
pipreqs=0.4.11=pypi_0
pluggy=1.0.0=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=py310h2bbff1b_0
publication=0.0.3=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.9=py310haa95532_0
pytest=7.1.3=pypi_0
python=3.10.6=hbb2ffb3_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=pypi_0
pytz=2022.4=pypi_0
pywin32=302=py310h2bbff1b_2
pyzmq=23.2.0=py310hd77b12b_0
requests=2.28.1=pypi_0
rfc3986=1.5.0=pypi_0
setuptools=63.4.1=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
sqlite=3.39.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
tornado=6.2=py310h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typeguard=2.13.3=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.12=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py310haa95532_2
xz=5.2.6=h8cc25b3_0
yarg=0.1.9=pypi_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.12=h8cc25b3_3
